deduplicate handleChange into parseFormChange
