+++ react-ui/src/tabs/TortoiseGenerationParams.tsx
@@ -1,4 +1,6 @@
-import { updateLocalStorageWithFunction } from "../hooks/useLocalStorage";
+import useLocalStorage, {
+  updateLocalStorageWithFunction,
+} from "../hooks/useLocalStorage";
 import router from "next/router";
 
 export type TortoiseGenerationParams = {
@@ -71,3 +73,6 @@ export const sendToBarkVoiceGeneration = (audio?: string) => {
   );
   router.push("/bark");
 };
+
+export const useTortoiseGenerationParams = () =>
+  useLocalStorage<TortoiseGenerationParams>(tortoiseGenerationId, initialState);
