+++ react-ui/src/tabs/TortoiseResult.tsx
@@ -1,0 +1,10 @@
+import { GradioFile } from "../types/GradioFile";
+
+
+export type TortoiseResult = {
+  audio: GradioFile;
+  image: string;
+  seed: string;
+  bundle_name: string;
+  metadata: {};
+};
